[#54] refactor: 랜딩 페이지 리팩토링
